Tests/LibWeb/Text/expected/float-shrink-to-fit-width-after-containing-block-height-change.txt

before toggle: 100x50
after toggle:  100x100
reference:     200x100
FAIL: toggled float differs from from-scratch layout of the same state
